third_party/WebKit/LayoutTests/media/video-controls-download-button-not-displayed-local.html

<!DOCTYPE HTML>
<title>Download button should not appear for local files.</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="media-controls.js"></script>
<video controls></video>
<script>
async_test(function(t) {
  var video = document.querySelector("video");
  video.src = "content/test.ogv";
  var controlID = "-internal-media-controls-download-button";
  var downloadButton = mediaControlsElement(internals.shadowRoot(video).firstChild, controlID);
  video.onloadeddata = t.step_func_done(function() {
      assert_equals(getComputedStyle(downloadButton).display, "none");
  });
});
</script>
